+++ common/src/main/java/net/bettercombat/client/compat/ClientCompat.java
@@ -1,0 +1,7 @@
+package net.bettercombat.client.compat;
+
+public class ClientCompat {
+    public static void init() {
+        FirstPersonAnimationCompatibility.setup();
+    }
+}