+++ common/src/main/java/net/bettercombat/client/compat/PehkuiHelper.java
@@ -1,4 +1,4 @@
-package net.bettercombat.compatibility;
+package net.bettercombat.client.compat;
 
 import net.bettercombat.Platform;
 import net.bettercombat.api.client.AttackRangeExtensions;
@@ -9,7 +9,6 @@ import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Map;
-import java.util.function.Function;
 
 public class PehkuiHelper {
     public static Identifier scaleId = new Identifier("pehkui", "entity_reach");