+++ common/src/main/java/net/bettercombat/compatibility/CompatibilityFlags.java
@@ -1,37 +1,15 @@
 package net.bettercombat.compatibility;
 
 import net.bettercombat.Platform;
-
-import java.util.function.Supplier;
+import net.bettercombat.client.compat.PehkuiHelper;
 
 public class CompatibilityFlags {
-    public static boolean firstPersonRender() {
-        return firstPersonRender.get();
-    }
-    public static Supplier<Boolean> firstPersonRender = () -> { return true; };
     public static boolean usePehkui = false;
 
     public static void initialize() {
-        if (doesClassExist("dev.tr7zw.firstperson.FirstPersonModelCore")) {
-            firstPersonRender = FirstPersonModelHelper.isDisabled();
-        }
         if (Platform.isModLoaded("pehkui")) {
             usePehkui = true;
             PehkuiHelper.load();
         }
-    }
-
-    /**
-     * Checks if a class exists or not
-     * @param name
-     * @return
-     */
-    protected static boolean doesClassExist(String name) {
-        try {
-            if(Class.forName(name) != null) {
-                return true;
-            }
-        } catch (ClassNotFoundException e) {}
-        return false;
     }
 }