+++ common/src/main/java/net/bettercombat/config/TriStateAuto.java
@@ -1,0 +1,5 @@
+package net.bettercombat.config;
+
+public enum TriStateAuto {
+    YES, NO, AUTO
+}