+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Adjust API requirement to support Forge and NeoForge simultaneously #414
 - Fix weapon attributes with zero attacks (resulting due to conditions) crashing #472
 - Backport weapon registry compression #516
+- Backport updated first person view mod support #531 #540
 
 # 1.8.6